build output to sep module
